separate schedule components
