+++ client-course-schedulizer/src/components/reuseables/Schedule/scheduleComponents/index.ts
@@ -1,1 +1,3 @@
+export * from "./LeftTimeAxis";
 export * from "./ScheduleColumn";
+export * from "./ScheduleHeader";
